Further work on portfolio tracking

+++ src/Controller/AccountController.php
@@ -6,6 +6,7 @@ use NftPortfolioTracker\Entity\Account;
 use NftPortfolioTracker\Event\Account\AccountAddedEvent;
 use NftPortfolioTracker\Event\Account\AccountDeletedEvent;
 use NftPortfolioTracker\Event\Account\AccountUpdatedEvent;
+use NftPortfolioTracker\Repository\AccountBalanceRepository;
 use NftPortfolioTracker\Repository\AccountRepository;
 use Psr\EventDispatcher\EventDispatcherInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -19,11 +20,13 @@ class AccountController extends AbstractController
 {
     private EventDispatcherInterface $eventDispatcher;
     private AccountRepository $accountRepository;
+    private AccountBalanceRepository $accountBalanceRepository;
 
-    public function __construct(EventDispatcherInterface $eventDispatcher, AccountRepository $accountRepository)
+    public function __construct(EventDispatcherInterface $eventDispatcher, AccountRepository $accountRepository, AccountBalanceRepository $accountBalanceRepository)
     {
         $this->eventDispatcher = $eventDispatcher;
         $this->accountRepository = $accountRepository;
+        $this->accountBalanceRepository = $accountBalanceRepository;
     }
 
     /**
@@ -34,8 +37,40 @@ class AccountController extends AbstractController
         $accounts = $this->accountRepository->findAll();
 
         return $this->render('account/index.html.twig', [
-            'controller_name' => 'AccountController',
             'accounts' => $accounts,
+        ]);
+    }
+
+    /**
+     * @Route("/account/{address}", name="account_single")
+     */
+    public function single(string $address, Request $request): Response
+    {
+        if ($address === 'create') {
+            return $this->create($request);
+        }
+
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+
+        if (!$this->isGranted('ROLE_ADMIN') && $this->getUser()->getAccount()->getAddress() !== $address) {
+            $this->createAccessDeniedException('Trying to access another account');
+        }
+
+        $overallBalance = $this->accountBalanceRepository->findOneBy(['account' => $address, 'projectName' => null]);
+
+        $balances = $this->accountBalanceRepository
+            ->createQueryBuilder('balance')
+            ->select('balance')
+            ->where('balance.projectName is not null')
+            ->andWhere('balance.account = :account')
+            ->setParameter('account', $address)
+            ->orderBy('balance.balance', 'asc')
+            ->getQuery()->getResult();
+
+        array_unshift($balances, $overallBalance);
+
+        return $this->render('account/single.html.twig', [
+            'balances' => $balances,
         ]);
     }
 
@@ -71,26 +106,29 @@ class AccountController extends AbstractController
         }
 
         return $this->render('account/create.html.twig', [
-            'controller_name' => 'AccountController',
             'form' => $form->createView()
         ]);
     }
 
     /**
-     * @Route("/account/delete/{address}", name="account_delete")
+     * @Route("/account/{address}/delete", name="account_delete")
      */
     public function delete(string $address): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $this->eventDispatcher->dispatch(AccountDeletedEvent::create($address));
 
         return $this->redirectToRoute('account');
     }
 
     /**
-     * @Route("/account/refresh/{address}", name="account_refresh")
+     * @Route("/account/{address}/refresh", name="account_refresh")
      */
     public function refresh(string $address): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $this->eventDispatcher->dispatch(AccountUpdatedEvent::create($address, []));
 
         return $this->redirectToRoute('account');
